pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "finops-ai"
dynamic = ["version"]
description = "AI-powered multi-cloud cost optimization for humans"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "FinOps AI Contributors"},
]
keywords = ["finops", "cloud", "cost-optimization", "azure", "aws", "gcp", "ai", "ml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "tabulate>=0.9.0",
    "requests>=2.31.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.12.0",
    "azure-mgmt-compute>=29.0.0",
    "azure-mgmt-resource>=21.1.0",
    "azure-mgmt-network>=23.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-web>=7.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "azure-storage-blob>=12.14.0",
]
aws = [
    "boto3>=1.28.0",
]
gcp = [
    "google-cloud-compute>=1.14.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-monitoring>=2.16.0",
    "google-auth>=2.22.0",
]
ml = [
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "prophet>=1.1.4",
]
all = [
    "finops-ai[azure,aws,gcp,ml]",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "build>=1.0.0",
]

[project.scripts]
finops-ai = "finops_ai.cli:cli"

[project.urls]
Homepage = "https://github.com/finops-ai/finops-ai"
Documentation = "https://finops-ai.readthedocs.io"
Repository = "https://github.com/finops-ai/finops-ai"
Issues = "https://github.com/finops-ai/finops-ai/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools_scm]
fallback_version = "0.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring real cloud credentials",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.coverage.run]
source = ["src/finops_ai"]

[tool.coverage.report]
show_missing = true
fail_under = 80
